+++ connect/lib/core/src/main/scala/io/iohk/atala/connect/core/service/ConnectionServiceImpl.scala
@@ -174,6 +174,18 @@ private class ConnectionServiceImpl(
       case Some(value) => ZIO.succeed(value)
     }
 
+  override def markConnectionInvitationExpired(
+      recordId: UUID
+  ): ZIO[WalletAccessContext, ConnectionServiceError, ConnectionRecord] =
+    updateConnectionProtocolState(
+      recordId,
+      ProtocolState.InvitationGenerated,
+      ProtocolState.InvitationExpired
+    ).flatMap {
+      case None        => ZIO.fail(RecordIdNotFound(recordId))
+      case Some(value) => ZIO.succeed(value)
+    }
+
   override def receiveConnectionRequest(
       request: ConnectionRequest
   ): ZIO[WalletAccessContext, ConnectionServiceError, ConnectionRecord] =