+++ prism-agent/service/server/src/main/scala/io/iohk/atala/agent/server/config/AppConfig.scala
@@ -40,7 +40,8 @@ final case class ConnectConfig(
     database: DatabaseConfig,
     connectBgJobRecordsLimit: Int,
     connectBgJobRecurrenceDelay: Duration,
-    connectBgJobProcessingParallelism: Int
+    connectBgJobProcessingParallelism: Int,
+    connectInvitationExpiry: Duration,
 )
 
 final case class PrismNodeConfig(service: GrpcServiceConfig)